Extract DomStorageNodeBuilder
