+++ config-model/src/main/java/com/yahoo/vespa/model/content/StorageNode.java
@@ -2,21 +2,16 @@
 package com.yahoo.vespa.model.content;
 
 import com.yahoo.config.model.api.ModelContext;
-import com.yahoo.config.model.deploy.DeployState;
-import com.yahoo.config.model.producer.TreeConfigProducer;
 import com.yahoo.vespa.config.content.StorFilestorConfig;
 import com.yahoo.vespa.config.content.core.StorServerConfig;
 import com.yahoo.vespa.defaults.Defaults;
 import com.yahoo.vespa.model.application.validation.RestartConfigs;
-import com.yahoo.vespa.model.builder.xml.dom.ModelElement;
-import com.yahoo.vespa.model.builder.xml.dom.VespaDomBuilder;
 import com.yahoo.vespa.model.content.engines.PersistenceEngine;
 import com.yahoo.vespa.model.content.engines.ProtonProvider;
 import com.yahoo.vespa.model.content.storagecluster.StorageCluster;
-import org.w3c.dom.Element;
-import java.util.Optional;
 
-import static java.util.logging.Level.WARNING;
+import java.util.Objects;
+import java.util.Optional;
 
 /**
  * Class to provide config related to a specific storage node.
@@ -30,20 +25,8 @@ public class StorageNode extends ContentNode implements StorServerConfig.Produce
     private final boolean retired;
     private final StorageCluster cluster;
 
-    public static class Builder extends VespaDomBuilder.DomConfigProducerBuilder<StorageNode, StorageNode> {
-
-        @Override
-        protected StorageNode doBuild(DeployState deployState, TreeConfigProducer<StorageNode> ancestor, Element producerSpec) {
-            ModelElement e = new ModelElement(producerSpec);
-            Double capacity = e.doubleAttribute("capacity");
-            if (capacity != null)
-                deployState.getDeployLogger().logApplicationPackage(WARNING, "'capacity' is deprecated, see https://docs.vespa.ai/en/reference/services-content#node");
-            return new StorageNode(deployState.getProperties(), (StorageCluster)ancestor, capacity, e.integerAttribute("distribution-key"), false);
-        }
-
-    }
-
-    StorageNode(ModelContext.Properties properties, StorageCluster cluster, Double capacity, int distributionKey, boolean retired) {
+    public StorageNode(ModelContext.Properties properties, StorageCluster cluster,
+                       Double capacity, int distributionKey, boolean retired) {
         super(properties.featureFlags(), cluster, cluster.getClusterName(),
               rootFolder + cluster.getClusterName() + "/storage/" + distributionKey,
               distributionKey);
@@ -60,11 +43,7 @@ public class StorageNode extends ContentNode implements StorServerConfig.Produce
     }
 
     public double getCapacity() {
-        if (capacity != null) {
-            return capacity;
-        } else {
-            return 1.0;
-        }
+        return Objects.requireNonNullElse(capacity, 1.0);
     }
 
     /** Whether this node is configured as retired, which means all content should migrate off the node */
